Finally got the confidence interval to work
After a good deal of tweaking I have been able to get the existing implementation of the bootstrap confidence intervals to work, even if some of the details of its functioning still elude me. For one, I cannot find an explanation on what a "vector of proportions" actually is. I'll have to talk to Prof. Hartlaub about it.