Implemented the other types of interval
I've implemented and tested the various options for bootstrap confidence interval, but there is a bit of clunkiness on the user's part, as it is not clear what the available options are. In particular, I tripped myself up by including an extra space in the string name, and the program wasn't running correctly.