I have properly implemented and tested the k-sample bootstrap
Previously, I was not sure how to test, or to properly notate my code, and now I have done so. I do know there are a couple of gaps to fill in the case of unbalanced samples, but so far the formula works for producing intervals at least for linear models. Further tests will be needed.